feat: add MessagingAuth component for authentication handling in messaging
